+++ activitysampling-application/src/main/java/de/muspellheim/activitysampling/application/shared/Configuration.java
@@ -1,19 +1,30 @@
-package de.muspellheim.activitysampling.application;
+package de.muspellheim.activitysampling.application.shared;
 
-class Configuration {
-  static final Configuration INSTANCE = new Configuration();
+import java.time.*;
+
+public class Configuration {
+  public static final Configuration INSTANCE = new Configuration();
 
   private static final String PROP_LOG_FILE = "activitySampling.logFile";
   private static final String DEFAULT_LOG_FILE =
       System.getProperty("user.home") + System.lineSeparator() + "activity-log.csv";
 
+  private Clock clock = Clock.systemDefaultZone();
   private String logFile;
 
   private Configuration() {
     logFile = System.getProperty(PROP_LOG_FILE, DEFAULT_LOG_FILE);
   }
 
-  String getLogFile() {
+  public Clock getClock() {
+    return clock;
+  }
+
+  public void setClock(Clock clock) {
+    this.clock = clock;
+  }
+
+  public String getLogFile() {
     return logFile;
   }