+++ activitysampling-application/src/test/java/de/muspellheim/activitysampling/application/uat/TimesheetTests.java
@@ -1,0 +1,22 @@
+package de.muspellheim.activitysampling.application.uat;
+
+import org.junit.jupiter.api.*;
+
+class TimesheetTests {
+  @BeforeEach
+  void init() {
+    SystemUnderTest.INSTANCE.reset();
+  }
+
+  @Test
+  @Disabled
+  void mainScenario() {
+    // TODO implement acceptance test for timesheet
+    /*
+    var timesheetFixture = new TimesheetFixture();
+    var timesheetEntriesFixture = new TimesheetEntriesFixture();
+
+    activitySamplingFixture.now(Instant.parse("2022-11-16T17:05:00Z"));
+    */
+  }
+}