+++ activitysampling-domain/src/main/java/de/muspellheim/activitysampling/domain/Timesheet.java
@@ -3,4 +3,4 @@ package de.muspellheim.activitysampling.domain;
 import java.time.*;
 import java.util.*;
 
-public record Timesheet(List<TimesheetByDay> workingDays, Duration total) {}
+public record Timesheet(List<TimesheetEntry> entries, Duration total) {}